Unshare mount namespace before doing libguestfs
I'm seeing a bug where qemu spawned by guestmount sometimes spins at
100% CPU (presumably the agent is consuming the core).  I suspect
what's happening is something like udisks is getting a mount
notification and jumping in and doing a statvfs() before guestmount
is ready.

Regardless of the source of this bug, it's just better to use a
private mount namespace as it ensures that if the process exits, the
mount should be cleaned up.