installer: Require output dir specification
Previously we'd hardcoded images/<partial checksum>, but I'm finding
that what I really want to do is stage my choice of images, and then
put them into place manually.  Or if not manually, in a more flexible
way that can be encoded into this script.

So this change then just forces callers to specify an output
directory.