taskbase: Use values from ini, override with json
Certain required variables in ini files like ref were unused in
tree composes.  These fixes result in using setting a rule that
if the value is in the ini file, use it but it can be overriden
by input from a json file.  The ini variable is then added to a
temporary json file that is stored in a tempdir and consumed by
'rpm-ostree compose tree'.

Along with adding the ini content to the json file, a function
was also added to detect if the json file has an include
statement.  If so, it will inherit the included json file
content and merge the two with the following rules:

    * if the key value doesn't exist in the json, add it
    * if the key value is a list and does exist, merge and
      remove duplicates.

And finally, to make processing for 'rpm-ostree tree compose'
easier, the detected repo files are copied to the temp dir.