Restart services on failure.
Failures are common with QPID: network hiccups, planned
outages, gremlins.

Much as I wanted to leave this as a site policy decision,
there doesn't seem to be a simple systemctl-ish way to do so.
See systemd.service(5) for restart directive, and perhaps
rhbz#756787 for hints on how to use /etc/systemd for
local overrides.

On kinit.service, set RemainAfterExit=yes. This serves as
a flag for build-monitor.service