installer: Support injecting %post scripts
This is a horrific hack, I admit.  Basically as an OS builder
that requires subscription management, we need a way to inject
%post bits to explicitly deconfigure the system.

For example:

rm -f /etc/ostree/remotes.d/*.conf
echo 'unconfigured-state=This system is not registered; You can use subscription-manager to register.' >> $(ostree admin --print-current-dir).origin

A perhaps better alternative would be a variant of ostreesetup that
unset the remote configuration when done.  But that would require a
new pykickstart...