treecompose: add ability to output rhev and vsphere images
Requires new imagefactory-plugin deps and versions in spec.

Treecompose will now output multiple image flavors.  They are:

    * kvm (qcow2), raw, rhev (ova), vsphere (ova)

The above images will be the default outputs. If you pass a -i or
--images, you can override the ouputs.  For each image desired, you
pass an additional -i.  Treecompose will always output a qcow2.

For now, treecompose will require the default oz image_type to be raw
because imagefactory (for now) requires this for input on image
conversation.  When that is fixed, we then make raw no longer required.

Additionally, the -c option is no longer required. If not provided,
we look for config.ini.

And finally, there a some new exit conditions for treecompose.  We now
check the oz config for the image_type and a minimum RAM recommendation.
If those are not met, the script will exit with a stdout error message.
This is done so that consumer tools like koji/brew have the error at the
ready.