imagefactory: Fix bug OZ config memory parsing
This patch fixes a bug where if the memory option does not
exist in the libvirt section of the oz.cfg, it produced a
traceback.  We now check to see if it exists, before attempting
to work with it.  Additionally, we now use the ConfigParser method
to process the oz configuration file, as is more inline with the
rest of -toolbox.