Rework output handling, remove outputdir/ostree_repo config.ini settings
This kept causing bugs because the option of inheriting it from
os.getcwd() or explicitly specifying it caused us to flip between
making one or the other case buggy.

With this patch, outputdir is always os.getcwd().  The ostree repo
is always found at ${outputdir}/repo.

For generating images, one must specify --outputdir.  The system
will error out if they already exist (for release engineering use),
but as a convenience for local development, a new --overwrite
argument is provided.