imagefactory: Add ability to create hyperv and vagrant images
We can now output hyperv images and two forms of vagrant images.  For
hyper-v, we can output a Gen1 image that can be imported into a hyper-v
Gen1 VM.  The image lacks the OVA tagging and format because MS deviates
from the standrard as we know it.  For vagrant, we can now produce a
libvirt-based image and a virtualbox-base image.

* imagefactory:
    - Addition of -i hyperv -i vagrant-libvirt -i vagrant-virtualbox
    - Vagrant images require an additional kickstart which can be
      referred to with --vkickstart
    - Broke out several new functions from the create() to efficiently
      deal with building two base images when vagrant images are
      desired. i.e. generateOVA(), formatKS()