liveimage.py Add ability to create PXE Live Images
This patch adds the ability to create atomic PXE to
live images images.  It builds an atomic RAW disk
image and then uses live media-creator inside a native
container to output four items:

    * kernel
    * initrd
    * squashfs image
    * sample PXE configuration

Other changes in this patch include a Class refactoring
which made inclusion of the liveimage.py easier.  There is
also a small change to the Makefile so that the liveimage
file is installed.