s/print/log() and delete a lot of debugging prints
I am hitting an error with a compose.  It is greatly obscured by the
fact that when piping output to a log file, because stdout is
buffered, you'll only see it at the end.

Fix this by introducing a log() function that flushes stdout.

Also delete a lot of debugging prints.

We still need to be even less verbose than this, but this patch takes
us closer.