Add docker-image command, use it for installer/liveimage
I've ran into some really frustrating issues doing 'yum install'
inside 'docker build'.  A major one is that it's a tremendous pain in
the ass to try to teach it to use local package caches, or even just
to honor $http_proxy.

(Also, setting $http_proxy breaks /usr/bin/docker; unset it before invocation)

This patch then adds a 'docker-image' command that uses yum on the
host to make an image as a tarball (with no base) and loads it.

I plan to use this for other things, but it's obviously useful for
-toolbox's internals too.