Run "ksflatten" inside a Docker container too
I run RHEL7 Workstation, and the servers I run use RHEL7 Server.
However, we want to be able to do Fedora builds from them.  Having the
"ksflatten" command on the host has been a cause of incredible amounts
of pain, as it has to know about the target system.

Fix this by creating a container that has just pykickstart (+deps)
assembled using the target repositories, just like we do for Lorax.

This is the start of a more generic infrastructure for host-managed
releng tasks using Docker.