installer: Drop the imagefactory-based approach
It was useful to do both for evaluation...but containers are faster,
easier to debug.  Also, Fedora/CentOS/RHEL rel-eng all do lorax on the
host or in mock, not in imagefactory, so we're deviating.

The code duplication was making things harder to fix.  It's there in
the git history if we need it again for some reason.