Grant all containers --privileged for easier data sharing
This obviously kind of sucks, but currently it's required to label
data accessed by non-privileged containers with
`svirt_sandbox_file_t`.

We could just make a temporary copy of the files and relabel or
something, but since every other container is `--privileged=true`,
let's just kick this can down the road a bit.

Clearly doing more generic rel-eng stuff with Docker where data is
stored on the host is going to run into this more.