imagefactory: Automatically substitute kickstart --url argument
The fedora-atomic kickstarts reference Koji-generated ostreesetup
URLs.  Whereas often, I want to generate a tree locally, then have the
cloud images use that.

If we're taking the viewpoint that official rel-eng for images uses
Koji, then the kickstarts are fine, and -toolbox can be more of a
developer and "single integration server" tool.