installer: Only remove origin config if it's file:///install/ostree
For interactive installs from `installer.iso`, we pull content from
`file:///install/ostree`.  However, it's not really useful for that to
be the origin, so we were removing it in `%post`.

But doing so breaks for the case where we have
https://github.com/rhinstaller/anaconda/commit/438e19b8fe0f9f8d87a1a003f634d455973c87e1
(Fedora 21+, soon to be CentOS) *and* we are including a remote config
in the tree content (CentOS).

Let's be more conservative here and only delete the remote if it's
from `file:///install/ostree`.

Note that apparently the <%text>$$</%text> thing is how one needs to
do escaping in Mako.