Fix treefile import defaulting logic
The first bug was that somehow params was enclosed in quotes, so the
if conditions were constant expressions (no, 'ref' is not contained in
the string 'params').

The second bug was that the defaulting needed to happen after recursing
through imports, since the recursive flattenjsoninclude code will skip
loading attributes that are already set.

Fixes #87
Closes: #89
Approved by: cgwalters