installer: Support propagating repos into installer
I'm not sure offhand why we weren't doing this by default.  I'm
working on Atomic Workstation and put a patched anaconda into my
override repo, but it wasn't being picked up.

It's just a lot simpler to support a mode where the treefile repos are
also used for the installer run.